// chpqa-app/src/app/api-services/chpqa-api/generated/model/replySubmissionGroups.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ReplySubmissionGroups { 
    id?: string | null;
    name?: string | null;
    submittedName?: string | null;
    groupCategory?: number | null;
    groupType?: number | null;
    status?: number | null;
    assessorStatus?: number | null;
    displayOrder?: number | null;
}